+++ internal/provider/template.go
@@ -118,7 +118,7 @@ func (t providerFileTemplate) Render(name string) (string, error) {
 	}{name, providerShortName(name)})
 }
 
-func (t providerTemplate) Render(providerName, exampleFile string, schema *tfjson.Schema) (string, error) {
+func (t providerTemplate) Render(providerName, renderedProviderName, exampleFile string, schema *tfjson.Schema) (string, error) {
 	schemaBuffer := bytes.NewBuffer(nil)
 	err := schemamd.Render(schema, schemaBuffer)
 	if err != nil {
@@ -144,6 +144,8 @@ func (t providerTemplate) Render(providerName, exampleFile string, schema *tfjso
 		ProviderShortName string
 
 		SchemaMarkdown string
+
+		RenderedProviderName string
 	}{
 		Description: schema.Block.Description,
 
@@ -154,10 +156,12 @@ func (t providerTemplate) Render(providerName, exampleFile string, schema *tfjso
 		ProviderShortName: providerShortName(providerName),
 
 		SchemaMarkdown: schemaComment + "\n" + schemaBuffer.String(),
+
+		RenderedProviderName: renderedProviderName,
 	})
 }
 
-func (t resourceTemplate) Render(name, providerName, typeName, exampleFile, importFile string, schema *tfjson.Schema) (string, error) {
+func (t resourceTemplate) Render(name, providerName, renderedProviderName, typeName, exampleFile, importFile string, schema *tfjson.Schema) (string, error) {
 	schemaBuffer := bytes.NewBuffer(nil)
 	err := schemamd.Render(schema, schemaBuffer)
 	if err != nil {
@@ -184,6 +188,8 @@ func (t resourceTemplate) Render(name, providerName, typeName, exampleFile, impo
 		ProviderShortName string
 
 		SchemaMarkdown string
+
+		RenderedProviderName string
 	}{
 		Type:        typeName,
 		Name:        name,
@@ -199,6 +205,8 @@ func (t resourceTemplate) Render(name, providerName, typeName, exampleFile, impo
 		ProviderShortName: providerShortName(providerName),
 
 		SchemaMarkdown: schemaComment + "\n" + schemaBuffer.String(),
+
+		RenderedProviderName: renderedProviderName,
 	})
 }