+++ CHANGELOG.md
@@ -5,6 +5,10 @@ BUG FIXES:
 * cmd/tfplugindocs: Allow single word resources to use templates ([#147](https://github.com/hashicorp/terraform-plugin-docs/pull/147)).
 * cmd/tfplugindocs: Pass in correct provider name for data-source and resource schema lookup when overidden with `rendered-provider-name` flag ([#148](https://github.com/hashicorp/terraform-plugin-docs/pull/148)).
 
+ENHANCEMENTS:
+
+* cmd/tfplugindocs: Expose `RenderedProviderName` to templates ([#149](https://github.com/hashicorp/terraform-plugin-docs/pull/149)).
+
 # 0.9.0 (June 1, 2022)
 
 NEW FEATURES: